Add client implementation
This adds a sse client to to this library as found in
github.com/andrewstuart/go-sse with MIT license to avoid
cluttering server and client implementations all over the place and to
prepare for future development here.

So this library is now essentially a merge between
github.com/JanBerktold/sse and github.com/andrewstuart/go-sse.